🎨 💄 投稿詳細画面や、新規投稿作成画面で前の画面に戻る必要がある場合のNavbar #1
表示させたいタイトルがある場合は、propsで渡す
各ページでimportするように
